clean up old code, cast to float, u

+++ flood_forecast/preprocessing/pytorch_loaders.py
@@ -95,7 +95,7 @@ class CSVTestLoader(CSVDataLoader):
             targ_rows = self.df.iloc[target_idx_start:self.forecast_total+target_idx_start]
             all_rows_orig = self.original_df.iloc[idx:self.forecast_total+target_idx_start]
             historical_rows = torch.from_numpy(historical_rows.to_numpy())
-            return historical_rows, all_rows_orig, target_idx_start
+            return historical_rows.float(), all_rows_orig, target_idx_start
 
     def __len__(self):
         return len(self.df.index)-self.forecast_history-self.forecast_total-1
